requirements.txt
django
requests
python-dotenv
djangorestframework
djangorestframework-simplejwt